+++ src/test/java/org/morphix/lang/function/RunnablesTest.java
@@ -14,12 +14,17 @@ package org.morphix.lang.function;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
+import java.lang.reflect.InvocationTargetException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Supplier;
 
 import org.junit.jupiter.api.Test;
+import org.morphix.lang.JavaObjects;
+import org.morphix.reflection.Constructors;
+import org.morphix.reflection.ReflectionException;
 
 /**
  * Test class for {@link Runnables}.
@@ -48,4 +53,11 @@ class RunnablesTest {
 		assertThat(list, equalTo(List.of(1, 2)));
 	}
 
+	@Test
+	void shouldThrowExceptionWhenTryingToInstantiateClass() {
+		ReflectionException reflectionException = assertThrows(ReflectionException.class, () -> Constructors.IgnoreAccess.newInstance(Runnables.class));
+		InvocationTargetException invocationTargetException = JavaObjects.cast(reflectionException.getCause());
+		UnsupportedOperationException unsupportedOperationException = JavaObjects.cast(invocationTargetException.getCause());
+		assertThat(unsupportedOperationException.getMessage(), equalTo(Constructors.MESSAGE_THIS_CLASS_SHOULD_NOT_BE_INSTANTIATED));
+	}
 }